+++ crates/ruma-client-api/src/error.rs
@@ -1,6 +1,6 @@
 //! Errors that can be sent from the homeserver.
 
-use std::{collections::BTreeMap, fmt, str::FromStr, sync::Arc};
+use std::{fmt, str::FromStr, sync::Arc};
 
 use as_variant::as_variant;
 use bytes::{BufMut, Bytes};
@@ -13,7 +13,7 @@ use ruma_common::{
             IntoHttpError, MatrixErrorBody,
         },
     },
-    serde::StringEnum,
+    serde::{JsonObject, StringEnum},
 };
 use serde::{Deserialize, Serialize};
 use serde_json::{Value as JsonValue, from_slice as from_json_slice};
@@ -69,13 +69,7 @@ pub enum ErrorKind {
     /// `M_BAD_STATUS`
     ///
     /// The application service returned a bad status.
-    BadStatus {
-        /// The HTTP status code of the response.
-        status: Option<http::StatusCode>,
-
-        /// The body of the response.
-        body: Option<String>,
-    },
+    BadStatus(BadStatusErrorData),
 
     /// `M_CANNOT_LEAVE_SERVER_NOTICE_ROOM`
     ///
@@ -137,11 +131,7 @@ pub enum ErrorKind {
     ///
     /// Forbidden access, e.g. joining a room without permission, failed login.
     #[non_exhaustive]
-    Forbidden {
-        /// The `WWW-Authenticate` header error message.
-        #[cfg(feature = "unstable-msc2967")]
-        authenticate: Option<AuthenticateError>,
-    },
+    Forbidden,
 
     /// `M_GUEST_ACCESS_FORBIDDEN`
     ///
@@ -153,10 +143,7 @@ pub enum ErrorKind {
     /// `M_INCOMPATIBLE_ROOM_VERSION`
     ///
     /// The client attempted to join a room that has a version the server does not support.
-    IncompatibleRoomVersion {
-        /// The room's version.
-        room_version: RoomVersionId,
-    },
+    IncompatibleRoomVersion(IncompatibleRoomVersionErrorData),
 
     /// `M_INVALID_PARAM`
     ///
@@ -189,10 +176,7 @@ pub enum ErrorKind {
     /// short period of time.
     ///
     /// [rate limiting]: https://spec.matrix.org/latest/client-server-api/#rate-limiting
-    LimitExceeded {
-        /// How long a client should wait before they can try again.
-        retry_after: Option<RetryAfter>,
-    },
+    LimitExceeded(LimitExceededErrorData),
 
     /// `M_MISSING_PARAM`
     ///
@@ -238,10 +222,7 @@ pub enum ErrorKind {
     /// The request cannot be completed because the homeserver has reached a resource limit imposed
     /// on it. For example, a homeserver held in a shared hosting environment may reach a resource
     /// limit if it starts using too much memory or disk space.
-    ResourceLimitExceeded {
-        /// A URI giving a contact method for the server administrator.
-        admin_contact: String,
-    },
+    ResourceLimitExceeded(ResourceLimitExceededErrorData),
 
     /// `M_ROOM_IN_USE`
     ///
@@ -353,14 +334,7 @@ pub enum ErrorKind {
     /// The [access or refresh token] specified was not recognized.
     ///
     /// [access or refresh token]: https://spec.matrix.org/latest/client-server-api/#client-authentication
-    UnknownToken {
-        /// If this is `true`, the client is in a "[soft logout]" state, i.e. the server requires
-        /// re-authentication but the session is not invalidated. The client can acquire a new
-        /// access token by specifying the device ID it is already using to the login API.
-        ///
-        /// [soft logout]: https://spec.matrix.org/latest/client-server-api/#soft-logout
-        soft_logout: bool,
-    },
+    UnknownToken(UnknownTokenErrorData),
 
     /// `M_UNRECOGNIZED`
     ///
@@ -420,31 +394,13 @@ pub enum ErrorKind {
     /// backup version.
     ///
     /// [room keys backup]: https://spec.matrix.org/latest/client-server-api/#server-side-key-backups
-    WrongRoomKeysVersion {
-        /// The currently active backup version.
-        current_version: Option<String>,
-    },
+    WrongRoomKeysVersion(WrongRoomKeysVersionErrorData),
 
     #[doc(hidden)]
-    _Custom { errcode: PrivOwnedStr, extra: Extra },
+    _Custom(CustomErrorKind),
 }
 
 impl ErrorKind {
-    /// Constructs an empty [`ErrorKind::Forbidden`] variant.
-    pub fn forbidden() -> Self {
-        Self::Forbidden {
-            #[cfg(feature = "unstable-msc2967")]
-            authenticate: None,
-        }
-    }
-
-    /// Constructs an [`ErrorKind::Forbidden`] variant with the given `WWW-Authenticate` header
-    /// error message.
-    #[cfg(feature = "unstable-msc2967")]
-    pub fn forbidden_with_authenticate(authenticate: AuthenticateError) -> Self {
-        Self::Forbidden { authenticate: Some(authenticate) }
-    }
-
     /// Get the [`ErrorCode`] for this `ErrorKind`.
     pub fn errcode(&self) -> ErrorCode {
         match self {
@@ -452,7 +408,7 @@ impl ErrorKind {
             ErrorKind::BadAlias => ErrorCode::BadAlias,
             ErrorKind::BadJson => ErrorCode::BadJson,
             ErrorKind::BadState => ErrorCode::BadState,
-            ErrorKind::BadStatus { .. } => ErrorCode::BadStatus,
+            ErrorKind::BadStatus(_) => ErrorCode::BadStatus,
             ErrorKind::CannotLeaveServerNoticeRoom => ErrorCode::CannotLeaveServerNoticeRoom,
             ErrorKind::CannotOverwriteMedia => ErrorCode::CannotOverwriteMedia,
             ErrorKind::CaptchaInvalid => ErrorCode::CaptchaInvalid,
@@ -463,14 +419,14 @@ impl ErrorKind {
             ErrorKind::ConnectionTimeout => ErrorCode::ConnectionTimeout,
             ErrorKind::DuplicateAnnotation => ErrorCode::DuplicateAnnotation,
             ErrorKind::Exclusive => ErrorCode::Exclusive,
-            ErrorKind::Forbidden { .. } => ErrorCode::Forbidden,
+            ErrorKind::Forbidden => ErrorCode::Forbidden,
             ErrorKind::GuestAccessForbidden => ErrorCode::GuestAccessForbidden,
-            ErrorKind::IncompatibleRoomVersion { .. } => ErrorCode::IncompatibleRoomVersion,
+            ErrorKind::IncompatibleRoomVersion(_) => ErrorCode::IncompatibleRoomVersion,
             ErrorKind::InvalidParam => ErrorCode::InvalidParam,
             ErrorKind::InvalidRoomState => ErrorCode::InvalidRoomState,
             ErrorKind::InvalidUsername => ErrorCode::InvalidUsername,
             ErrorKind::InviteBlocked => ErrorCode::InviteBlocked,
-            ErrorKind::LimitExceeded { .. } => ErrorCode::LimitExceeded,
+            ErrorKind::LimitExceeded(_) => ErrorCode::LimitExceeded,
             ErrorKind::MissingParam => ErrorCode::MissingParam,
             ErrorKind::MissingToken => ErrorCode::MissingToken,
             ErrorKind::NotFound => ErrorCode::NotFound,
@@ -478,7 +434,7 @@ impl ErrorKind {
             ErrorKind::NotInThread => ErrorCode::NotInThread,
             ErrorKind::NotJson => ErrorCode::NotJson,
             ErrorKind::NotYetUploaded => ErrorCode::NotYetUploaded,
-            ErrorKind::ResourceLimitExceeded { .. } => ErrorCode::ResourceLimitExceeded,
+            ErrorKind::ResourceLimitExceeded(_) => ErrorCode::ResourceLimitExceeded,
             ErrorKind::RoomInUse => ErrorCode::RoomInUse,
             ErrorKind::ServerNotTrusted => ErrorCode::ServerNotTrusted,
             ErrorKind::ThreepidAuthFailed => ErrorCode::ThreepidAuthFailed,
@@ -496,7 +452,7 @@ impl ErrorKind {
             ErrorKind::Unknown => ErrorCode::Unknown,
             #[cfg(feature = "unstable-msc4186")]
             ErrorKind::UnknownPos => ErrorCode::UnknownPos,
-            ErrorKind::UnknownToken { .. } => ErrorCode::UnknownToken,
+            ErrorKind::UnknownToken(_) => ErrorCode::UnknownToken,
             ErrorKind::Unrecognized => ErrorCode::Unrecognized,
             ErrorKind::UnsupportedRoomVersion => ErrorCode::UnsupportedRoomVersion,
             ErrorKind::UrlNotSet => ErrorCode::UrlNotSet,
@@ -505,15 +461,125 @@ impl ErrorKind {
             ErrorKind::UserLocked => ErrorCode::UserLocked,
             ErrorKind::UserSuspended => ErrorCode::UserSuspended,
             ErrorKind::WeakPassword => ErrorCode::WeakPassword,
-            ErrorKind::WrongRoomKeysVersion { .. } => ErrorCode::WrongRoomKeysVersion,
-            ErrorKind::_Custom { errcode, .. } => errcode.0.clone().into(),
+            ErrorKind::WrongRoomKeysVersion(_) => ErrorCode::WrongRoomKeysVersion,
+            ErrorKind::_Custom(CustomErrorKind { errcode, .. }) => errcode.as_str().into(),
         }
+    }
+
+    /// Get the JSON data for this `ErrorKind`, if it uses a custom error code.
+    pub fn custom_json_data(&self) -> Option<&JsonObject> {
+        as_variant!(self, Self::_Custom).map(|error_kind| &error_kind.data)
     }
 }
 
+/// Data for the `M_BAD_STATUS` [`ErrorKind`].
+#[derive(Clone, Debug, Default, PartialEq, Eq)]
+#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
+pub struct BadStatusErrorData {
+    /// The HTTP status code of the response.
+    pub status: Option<http::StatusCode>,
+
+    /// The body of the response.
+    pub body: Option<String>,
+}
+
+impl BadStatusErrorData {
+    /// Construct a new empty `BadStatusErrorData`.
+    pub fn new() -> Self {
+        Self::default()
+    }
+}
+
+/// Data for the `M_INCOMPATIBLE_ROOM_VERSION` [`ErrorKind`].
+#[derive(Clone, Debug, PartialEq, Eq)]
+#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
+pub struct IncompatibleRoomVersionErrorData {
+    /// The room's version.
+    pub room_version: RoomVersionId,
+}
+
+impl IncompatibleRoomVersionErrorData {
+    /// Construct a new `IncompatibleRoomVersionErrorData` with the given room version.
+    pub fn new(room_version: RoomVersionId) -> Self {
+        Self { room_version }
+    }
+}
+
+/// Data for the `M_LIMIT_EXCEEDED` [`ErrorKind`].
+#[derive(Clone, Debug, Default, PartialEq, Eq)]
+#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
+pub struct LimitExceededErrorData {
+    /// How long a client should wait before they can try again.
+    pub retry_after: Option<RetryAfter>,
+}
+
+impl LimitExceededErrorData {
+    /// Construct a new empty `LimitExceededErrorData`.
+    pub fn new() -> Self {
+        Self::default()
+    }
+}
+
+/// Data for the `M_RESOURCE_LIMIT_EXCEEDED` [`ErrorKind`].
+#[derive(Clone, Debug, PartialEq, Eq)]
+#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
+pub struct ResourceLimitExceededErrorData {
+    /// A URI giving a contact method for the server administrator.
+    pub admin_contact: String,
+}
+
+impl ResourceLimitExceededErrorData {
+    /// Construct a new `ResourceLimitExceededErrorData` with the given admin contact URI.
+    pub fn new(admin_contact: String) -> Self {
+        Self { admin_contact }
+    }
+}
+
+/// Data for the `M_UNKNOWN_TOKEN` [`ErrorKind`].
+#[derive(Clone, Debug, Default, PartialEq, Eq)]
+#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
+pub struct UnknownTokenErrorData {
+    /// If this is `true`, the client is in a "[soft logout]" state, i.e. the server requires
+    /// re-authentication but the session is not invalidated. The client can acquire a new
+    /// access token by specifying the device ID it is already using to the login API.
+    ///
+    /// [soft logout]: https://spec.matrix.org/latest/client-server-api/#soft-logout
+    pub soft_logout: bool,
+}
+
+impl UnknownTokenErrorData {
+    /// Construct a new `UnknownTokenErrorData` with `soft_logout` set to `false`.
+    pub fn new() -> Self {
+        Self::default()
+    }
+}
+
+/// Data for the `M_WRONG_ROOM_KEYS_VERSION` [`ErrorKind`].
+#[derive(Clone, Debug, PartialEq, Eq)]
+#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
+pub struct WrongRoomKeysVersionErrorData {
+    /// The currently active backup version.
+    pub current_version: String,
+}
+
+impl WrongRoomKeysVersionErrorData {
+    /// Construct a new `WrongRoomKeysVersionErrorData` with the given current active backup
+    /// version.
+    pub fn new(current_version: String) -> Self {
+        Self { current_version }
+    }
+}
+
+/// A custom error kind.
 #[doc(hidden)]
 #[derive(Clone, Debug, PartialEq, Eq)]
-pub struct Extra(BTreeMap<String, JsonValue>);
+pub struct CustomErrorKind {
+    /// The error code.
+    errcode: String,
+
+    /// The data for the error.
+    data: JsonObject,
+}
 
 /// The possible [error codes] defined in the Matrix spec.
 ///
@@ -964,24 +1030,16 @@ impl EndpointError for Error {
             Ok(mut standard_body) => {
                 let headers = response.headers();
 
-                match &mut standard_body.kind {
-                    #[cfg(feature = "unstable-msc2967")]
-                    ErrorKind::Forbidden { authenticate } => {
-                        *authenticate = headers
-                            .get(http::header::WWW_AUTHENTICATE)
-                            .and_then(|val| val.to_str().ok())
-                            .and_then(AuthenticateError::from_str);
+                if let ErrorKind::LimitExceeded(LimitExceededErrorData { retry_after }) =
+                    &mut standard_body.kind
+                {
+                    // The Retry-After header takes precedence over the retry_after_ms field in
+                    // the body.
+                    if let Some(Ok(retry_after_header)) =
+                        headers.get(http::header::RETRY_AFTER).map(RetryAfter::try_from)
+                    {
+                        *retry_after = Some(retry_after_header);
                     }
-                    ErrorKind::LimitExceeded { retry_after } => {
-                        // The Retry-After header takes precedence over the retry_after_ms field in
-                        // the body.
-                        if let Some(Ok(retry_after_header)) =
-                            headers.get(http::header::RETRY_AFTER).map(RetryAfter::try_from)
-                        {
-                            *retry_after = Some(retry_after_header);
-                        }
-                    }
-                    _ => {}
                 }
 
                 ErrorBody::Standard(standard_body)
@@ -1031,19 +1089,13 @@ impl OutgoingResponse for Error {
             .header(http::header::CONTENT_TYPE, ruma_common::http_headers::APPLICATION_JSON)
             .status(self.status_code);
 
-        #[allow(clippy::collapsible_match)]
-        if let Some(kind) = self.error_kind() {
-            match kind {
-                #[cfg(feature = "unstable-msc2967")]
-                ErrorKind::Forbidden { authenticate: Some(auth_error) } => {
-                    builder = builder.header(http::header::WWW_AUTHENTICATE, auth_error);
-                }
-                ErrorKind::LimitExceeded { retry_after: Some(retry_after) } => {
-                    let header_value = http::HeaderValue::try_from(retry_after)?;
-                    builder = builder.header(http::header::RETRY_AFTER, header_value);
-                }
-                _ => {}
-            }
+        // Add data in headers.
+        if let Some(ErrorKind::LimitExceeded(LimitExceededErrorData {
+            retry_after: Some(retry_after),
+        })) = self.error_kind()
+        {
+            let header_value = http::HeaderValue::try_from(retry_after)?;
+            builder = builder.header(http::header::RETRY_AFTER, header_value);
         }
 
         builder
@@ -1059,101 +1111,6 @@ impl OutgoingResponse for Error {
                 }
             })
             .map_err(Into::into)
-    }
-}
-
-/// Errors in the `WWW-Authenticate` header.
-///
-/// To construct this use `::from_str()`. To get its serialized form, use its
-/// `TryInto<http::HeaderValue>` implementation.
-#[cfg(feature = "unstable-msc2967")]
-#[derive(Clone, Debug, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum AuthenticateError {
-    /// insufficient_scope
-    ///
-    /// Encountered when authentication is handled by OpenID Connect and the current access token
-    /// isn't authorized for the proper scope for this request. It should be paired with a
-    /// `401` status code and a `M_FORBIDDEN` error.
-    InsufficientScope {
-        /// The new scope to request an authorization for.
-        scope: String,
-    },
-
-    #[doc(hidden)]
-    _Custom { errcode: PrivOwnedStr, attributes: AuthenticateAttrs },
-}
-
-#[cfg(feature = "unstable-msc2967")]
-#[doc(hidden)]
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct AuthenticateAttrs(BTreeMap<String, String>);
-
-#[cfg(feature = "unstable-msc2967")]
-impl AuthenticateError {
-    /// Construct an `AuthenticateError` from a string.
-    ///
-    /// Returns `None` if the string doesn't contain an error.
-    fn from_str(s: &str) -> Option<Self> {
-        if let Some(val) = s.strip_prefix("Bearer").map(str::trim) {
-            let mut errcode = None;
-            let mut attrs = BTreeMap::new();
-
-            // Split the attributes separated by commas and optionally spaces, then split the keys
-            // and the values, with the values optionally surrounded by double quotes.
-            for (key, value) in val
-                .split(',')
-                .filter_map(|attr| attr.trim().split_once('='))
-                .map(|(key, value)| (key, value.trim_matches('"')))
-            {
-                if key == "error" {
-                    errcode = Some(value);
-                } else {
-                    attrs.insert(key.to_owned(), value.to_owned());
-                }
-            }
-
-            if let Some(errcode) = errcode {
-                let error = if let Some(scope) =
-                    attrs.get("scope").filter(|_| errcode == "insufficient_scope")
-                {
-                    AuthenticateError::InsufficientScope { scope: scope.to_owned() }
-                } else {
-                    AuthenticateError::_Custom {
-                        errcode: PrivOwnedStr(errcode.into()),
-                        attributes: AuthenticateAttrs(attrs),
-                    }
-                };
-
-                return Some(error);
-            }
-        }
-
-        None
-    }
-}
-
-#[cfg(feature = "unstable-msc2967")]
-impl TryFrom<&AuthenticateError> for http::HeaderValue {
-    type Error = http::header::InvalidHeaderValue;
-
-    fn try_from(error: &AuthenticateError) -> Result<Self, Self::Error> {
-        let s = match error {
-            AuthenticateError::InsufficientScope { scope } => {
-                format!("Bearer error=\"insufficient_scope\", scope=\"{scope}\"")
-            }
-            AuthenticateError::_Custom { errcode, attributes } => {
-                let mut s = format!("Bearer error=\"{}\"", errcode.0);
-
-                for (key, value) in attributes.0.iter() {
-                    s.push_str(&format!(", {key}=\"{value}\""));
-                }
-
-                s
-            }
-        };
-
-        s.try_into()
     }
 }
 
@@ -1217,7 +1174,10 @@ mod tests {
     };
     use web_time::{Duration, UNIX_EPOCH};
 
-    use super::{Error, ErrorBody, ErrorKind, RetryAfter, StandardErrorBody};
+    use super::{
+        Error, ErrorBody, ErrorKind, LimitExceededErrorData, RetryAfter, StandardErrorBody,
+        WrongRoomKeysVersionErrorData,
+    };
 
     #[test]
     fn deserialize_forbidden() {
@@ -1227,13 +1187,7 @@ mod tests {
         }))
         .unwrap();
 
-        assert_eq!(
-            deserialized.kind,
-            ErrorKind::Forbidden {
-                #[cfg(feature = "unstable-msc2967")]
-                authenticate: None
-            }
-        );
+        assert_eq!(deserialized.kind, ErrorKind::Forbidden,);
         assert_eq!(deserialized.message, "You are not authorized to ban users in this room.");
     }
 
@@ -1246,66 +1200,12 @@ mod tests {
         }))
         .expect("We should be able to deserialize a wrong room keys version error");
 
-        assert_matches!(deserialized.kind, ErrorKind::WrongRoomKeysVersion { current_version });
-        assert_eq!(current_version.as_deref(), Some("42"));
-        assert_eq!(deserialized.message, "Wrong backup version.");
-    }
-
-    #[cfg(feature = "unstable-msc2967")]
-    #[test]
-    fn custom_authenticate_error_sanity() {
-        use super::AuthenticateError;
-
-        let s = "Bearer error=\"custom_error\", misc=\"some content\"";
-
-        let error = AuthenticateError::from_str(s).unwrap();
-        let error_header = http::HeaderValue::try_from(&error).unwrap();
-
-        assert_eq!(error_header.to_str().unwrap(), s);
-    }
-
-    #[cfg(feature = "unstable-msc2967")]
-    #[test]
-    fn serialize_insufficient_scope() {
-        use super::AuthenticateError;
-
-        let error =
-            AuthenticateError::InsufficientScope { scope: "something_privileged".to_owned() };
-        let error_header = http::HeaderValue::try_from(&error).unwrap();
-
-        assert_eq!(
-            error_header.to_str().unwrap(),
-            "Bearer error=\"insufficient_scope\", scope=\"something_privileged\""
+        assert_matches!(
+            deserialized.kind,
+            ErrorKind::WrongRoomKeysVersion(WrongRoomKeysVersionErrorData { current_version })
         );
-    }
-
-    #[cfg(feature = "unstable-msc2967")]
-    #[test]
-    fn deserialize_insufficient_scope() {
-        use super::AuthenticateError;
-
-        let response = http::Response::builder()
-            .header(
-                http::header::WWW_AUTHENTICATE,
-                "Bearer error=\"insufficient_scope\", scope=\"something_privileged\"",
-            )
-            .status(http::StatusCode::UNAUTHORIZED)
-            .body(
-                serde_json::to_string(&json!({
-                    "errcode": "M_FORBIDDEN",
-                    "error": "Insufficient privilege",
-                }))
-                .unwrap(),
-            )
-            .unwrap();
-        let error = Error::from_http_response(response);
-
-        assert_eq!(error.status_code, http::StatusCode::UNAUTHORIZED);
-        assert_matches!(error.body, ErrorBody::Standard(StandardErrorBody { kind, message }));
-        assert_matches!(kind, ErrorKind::Forbidden { authenticate });
-        assert_eq!(message, "Insufficient privilege");
-        assert_matches!(authenticate, Some(AuthenticateError::InsufficientScope { scope }));
-        assert_eq!(scope, "something_privileged");
+        assert_eq!(current_version, "42");
+        assert_eq!(deserialized.message, "Wrong backup version.");
     }
 
     #[test]
@@ -1326,7 +1226,7 @@ mod tests {
         assert_matches!(
             error.body,
             ErrorBody::Standard(StandardErrorBody {
-                kind: ErrorKind::LimitExceeded { retry_after: None },
+                kind: ErrorKind::LimitExceeded(LimitExceededErrorData { retry_after: None }),
                 message
             })
         );
@@ -1352,7 +1252,9 @@ mod tests {
         assert_matches!(
             error.body,
             ErrorBody::Standard(StandardErrorBody {
-                kind: ErrorKind::LimitExceeded { retry_after: Some(retry_after) },
+                kind: ErrorKind::LimitExceeded(LimitExceededErrorData {
+                    retry_after: Some(retry_after)
+                }),
                 message
             })
         );
@@ -1380,7 +1282,9 @@ mod tests {
         assert_matches!(
             error.body,
             ErrorBody::Standard(StandardErrorBody {
-                kind: ErrorKind::LimitExceeded { retry_after: Some(retry_after) },
+                kind: ErrorKind::LimitExceeded(LimitExceededErrorData {
+                    retry_after: Some(retry_after)
+                }),
                 message
             })
         );
@@ -1408,7 +1312,9 @@ mod tests {
         assert_matches!(
             error.body,
             ErrorBody::Standard(StandardErrorBody {
-                kind: ErrorKind::LimitExceeded { retry_after: Some(retry_after) },
+                kind: ErrorKind::LimitExceeded(LimitExceededErrorData {
+                    retry_after: Some(retry_after)
+                }),
                 message
             })
         );
@@ -1437,7 +1343,9 @@ mod tests {
         assert_matches!(
             error.body,
             ErrorBody::Standard(StandardErrorBody {
-                kind: ErrorKind::LimitExceeded { retry_after: Some(retry_after) },
+                kind: ErrorKind::LimitExceeded(LimitExceededErrorData {
+                    retry_after: Some(retry_after)
+                }),
                 message
             })
         );
@@ -1451,7 +1359,7 @@ mod tests {
         let error = Error::new(
             http::StatusCode::TOO_MANY_REQUESTS,
             ErrorBody::Standard(StandardErrorBody {
-                kind: ErrorKind::LimitExceeded { retry_after: None },
+                kind: ErrorKind::LimitExceeded(LimitExceededErrorData { retry_after: None }),
                 message: "Too many requests".to_owned(),
             }),
         );
@@ -1476,9 +1384,9 @@ mod tests {
         let error = Error::new(
             http::StatusCode::TOO_MANY_REQUESTS,
             ErrorBody::Standard(StandardErrorBody {
-                kind: ErrorKind::LimitExceeded {
+                kind: ErrorKind::LimitExceeded(LimitExceededErrorData {
                     retry_after: Some(RetryAfter::Delay(Duration::from_secs(3))),
-                },
+                }),
                 message: "Too many requests".to_owned(),
             }),
         );
@@ -1505,11 +1413,11 @@ mod tests {
         let error = Error::new(
             http::StatusCode::TOO_MANY_REQUESTS,
             ErrorBody::Standard(StandardErrorBody {
-                kind: ErrorKind::LimitExceeded {
+                kind: ErrorKind::LimitExceeded(LimitExceededErrorData {
                     retry_after: Some(RetryAfter::DateTime(
                         UNIX_EPOCH + Duration::from_secs(1_431_704_061),
                     )),
-                },
+                }),
                 message: "Too many requests".to_owned(),
             }),
         );
@@ -1552,5 +1460,21 @@ mod tests {
                 "soft_logout": true,
             })
         );
+    }
+
+    #[test]
+    fn deserialize_custom_error_kind() {
+        let deserialized: StandardErrorBody = from_json_value(json!({
+            "errcode": "LOCAL_DEV_ERROR",
+            "error": "You are using the homeserver in local dev mode.",
+            "foo": "bar",
+        }))
+        .unwrap();
+
+        assert_eq!(deserialized.kind.errcode().as_str(), "LOCAL_DEV_ERROR");
+        assert_matches!(deserialized.kind.custom_json_data(), Some(json_data));
+        assert_matches!(json_data.get("foo"), Some(JsonValue::String(foo)));
+        assert_eq!(foo, "bar");
+        assert_eq!(deserialized.message, "You are using the homeserver in local dev mode.");
     }
 }